[fix] Extension 파일 구분
